--- a/src/FlowBasis/FlowBasis.Json/FlowBasis.Json.csproj
+++ b/src/FlowBasis/FlowBasis.Json/FlowBasis.Json.csproj
@@ -4,7 +4,7 @@
     <TargetFrameworks>netstandard2.0;net45</TargetFrameworks>
     <RootNamespace>FlowBasis.Json</RootNamespace>
     <AssemblyName>FlowBasis.Json</AssemblyName>
-    <Version>3.1.1</Version>
+    <Version>3.1.2</Version>
     <RepositoryUrl>https://github.com/calebdoise/flowbasis</RepositoryUrl>
     <Authors>Caleb Doise</Authors>
     <Company />
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
 </Project>