--- a/src/FlowBasis/FlowBasisExpressionsUnitTests/FlowBasisExpressionsUnitTests.csproj
+++ b/src/FlowBasis/FlowBasisExpressionsUnitTests/FlowBasisExpressionsUnitTests.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
 
@@ -11,13 +11,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.2.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.2.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.7.2" />
+    <PackageReference Include="MSTest.TestAdapter" Version="3.1.1" />
+    <PackageReference Include="MSTest.TestFramework" Version="3.1.1" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\FlowBasis.Expressions\FlowBasis.Expressions.csproj" />	
+    <ProjectReference Include="..\FlowBasis.Expressions\FlowBasis.Expressions.csproj" />
   </ItemGroup>
 
 </Project>