Hey Caleb, hope you are doing well.

I'm creating this PR in an effort to improve the performance of FlowBasis. I've noticed for some JSON payloads, FlowBasis will allocate about 10x the memory that Newtonsoft by itself. A lot of this seems to be due to the pre and post processing that FlowBasis does. The creation of IJObjectMappers in particular relies on a lot of reflection and allocates a lot of memory.

The main optimization is within `DefaultJObjectMapperProvider`. It was changed so that it now holds 2 ConcurrentDictionaries which are used to cache IJObjectMapper for types. If a custom mapper is dynamically added with `RegisterJObjectMapperForType`, then the cache is busted.

Another micro optimization was made where `JsonSerializationService` no longer allocates a new `Newtonsoft.Json.JsonSerializer` on every call to `Stringify`. It now holds onto the serializer as a private field.

# Benchmarking Results
```
BenchmarkDotNet v0.13.7, Windows 10 (10.0.19045.3448/22H2/2022Update)
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 7.0.401
  [Host]     : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2

```
|  Method |    Mean |    Error |   StdDev | Ratio | RatioSD |        Gen0 |       Gen1 |      Gen2 | Allocated | Alloc Ratio |
|-------- |--------:|---------:|---------:|------:|--------:|------------:|-----------:|----------:|----------:|------------:|
| OldCode | 4.312 s | 0.0823 s | 0.1923 s |  1.00 |    0.00 | 153000.0000 | 33000.0000 | 5000.0000 | 1199.6 MB |        1.00 |
| NewCode | 2.731 s | 0.0535 s | 0.0696 s |  0.64 |    0.03 |  67000.0000 | 27000.0000 | 4000.0000 | 518.17 MB |        0.43 |

The new code was benchmarked against the old code. The benchmark consisted of the following code

```
[Benchmark(Baseline = true)]
public IList<LargeProprietaryApiEndpointPayload> OldCode()
{
    var json = OldFlowBasis.JsonSerializers.Default.Stringify(this.payload);
    return OldFlowBasis.JsonSerializers.Default.Parse<IList<LargeProprietaryApiEndpointPayload>>(json);
}

[Benchmark]
public IList<LargeProprietaryApiEndpointPayload> NewCode()
{
    var json = NewFlowBasis.JsonSerializers.Default.Stringify(this.payload);
    return NewFlowBasis.JsonSerializers.Default.Parse<IList<LargeProprietaryApiEndpointPayload>>(json);
}
```

# Other Changes
* Bumped library version to 3.1.2
* Update Newtonsoft.Json from 13.0.1 to 13.0.3
* All .net core unit tests were updated to .net6.0
* All .net core unit tests had the following dependencies updated
    * Microsoft.NET.Test.Sdk
    *  MSTest.TestAdapter
    * MSTest.TestFramework
  * Updated various other private class fields to be `readonly`